requirements.txt
botbuilder-core>=4.7.0
aiohttp
botbuilder-dialogs>=4.7.0